requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-environ==0.11.2

pillow==9.3.0
psycopg2==2.9.5
django-ckeditor==6.3.2
django-storages==1.13.1

secure-smtplib==0.1.1

django-cloudinary-storage==0.3.0


gunicorn==20.1.0
whitenoise==6.5.0
dj-database-url==2.0.0